consortium-v2: support shadow fork with larger validator set
When shadow fork with larger validator set than the old one, the logic to get
checkpoint header is called twice in the same epoch. This makes the snapshot's
validator list updated twice and roll back the effect of shadow fork. This
commit prevents the snapshot's validator list from be updated twice in the same
epoch.