consortium-v2: upgrade Axie proxy contract (#462)
* params/config: add Aaron hardfork definition

* consortium-v2: upgrade Axie and Land transparent proxy code

* consortium-v2: upgrade Axie transparent proxy code

* consortium-v2: upgrade Land transparent proxy code

* consortium-v2: fix the code type to hextil bytes and update unit test

* consortium-v2: add block producer bit set (#2)

Block producer set is a subset of validator candidate set, but currently, both set are stored in plain, which costs unnecessary resource. This commit introduces BlockProducersBitSet (uint64) as an encoded representation of block producer list, indexing to validator candidate list. The original block producer set can be reconstructed from the bit set and validator candidate set. As a result, instead of storing a list of block producers in header extra, only a single bit set of type uint64 is stored.

* consortium-v2: add block producer bit set

* consortium-v2: add more check in verify cascading and remove bit set in snapshot

* consortium-v2: rename finality vote bit set to bit set

* consortium-v2: update unit test for header extra data encoding

* consortium-v2: add unit test for encode and decode validator bit set

* consortium-v2: fix nil checkpoint validator at epoch non-period block

* consortium-v2: add rlp optional tag to extradata encoding

* consortium-v2: add unit test decode rlp extra data

* consortium-v2: fix verify extra data and refactor code

* consortium-v2: add bit set check in verify extra data